requirements.txt
# Core Dagster
dagster==1.11.13
dagster-webserver==1.11.13
dagster-postgres

# DBT (installed separately, not via dagster-dbt)
dbt-trino==1.7.0

# Config & Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0

# Database
psycopg2-binary>=2.9.0

# AWS/S3
boto3>=1.28.0

# API Clients
requests>=2.31.0
ShopifyAPI>=12.0.0
tenacity>=8.2.3

pyarrow>=14.0.0
klaviyo_api